add ability to run several qemu with mounted filestore in test recipes; extract ssh client to lib from qemu recipe
