+++ cloud/filestore/tests/recipes/vhost-endpoint.inc
@@ -13,6 +13,10 @@ IF (FILESTORE_SHARD_COUNT)
     SET_APPEND(RECIPE_ARGS --shard-count $FILESTORE_SHARD_COUNT)
 ENDIF()
 
+IF (FILESTORE_VHOST_ENDPOINT_COUNT)
+    SET_APPEND(RECIPE_ARGS --endpoint-count $FILESTORE_VHOST_ENDPOINT_COUNT)
+ENDIF()
+
 USE_RECIPE(
     cloud/filestore/tests/recipes/vhost-endpoint/vhost-endpoint-recipe
     ${RECIPE_ARGS}
